skillswapp/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
firebase-admin==6.2.0
openai==0.28.1
Werkzeug==2.3.7
python-dotenv==1.0.0 